backend/requirements.txt
fastapi
uvicorn
numpy
faiss-cpu
python-dotenv
pydantic
sentence-transformers

